feat(like): forbid removing invalid like

+++ app/controllers/likes_controller.rb
@@ -13,7 +13,11 @@ class LikesController < ApplicationController
     fact_id = params[:fact_id]
     user = Current.user
     fact = user.user_likes_cat_facts.find_by(fact_id: fact_id)
-    fact.destroy
-    redirect_back_or_to root_url
+    if fact != nil
+      fact.destroy
+      redirect_back_or_to root_url
+    else
+      render file: "#{Rails.root}/public/400.html", layout: false, status: :not_found
+    end
   end
 end
